+++ src/Patch/ExecBlockFlags.h
@@ -15,37 +15,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifndef INSTMETADATA_H
-#define INSTMETADATA_H
+#ifndef ExecBlockFlags_H
+#define ExecBlockFlags_H
+
+#include <memory>
+#include <vector>
 
 #include "llvm/MC/MCInst.h"
+#include "llvm/MC/MCInstrInfo.h"
+#include "llvm/MC/MCRegisterInfo.h"
+#include "Platform.h"
 
-#include "Utility/InstAnalysis_prive.h"
-
-#include "State.h"
+#if defined(QBDI_ARCH_X86_64) || defined(QBDI_ARCH_X86)
+#include "Patch/X86_64/ExecBlockFlags_X86_64.h"
+#elif defined(QBDI_ARCH_ARM)
+#include "Patch/ARM/ExecBlockFlags_ARM.h"
+#endif
 
 namespace QBDI {
 
-class InstMetadata {
-public:
-    llvm::MCInst inst;
-    rword address;
-    uint32_t instSize;
-    uint32_t patchSize;
-    bool modifyPC;
-    bool merge;
-    uint8_t execblockFlags;
-    mutable InstAnalysisPtr analysis;
+uint8_t getExecBlockFlags(const llvm::MCInst& inst, const llvm::MCInstrInfo* MCII, const llvm::MCRegisterInfo* MRI);
 
-    inline rword endAddress() const {
-        return address + instSize;
-    }
-
-    inline InstMetadata lightCopy() const {
-        return {inst, address, instSize, patchSize, modifyPC, merge, execblockFlags, nullptr};
-    }
-};
+extern const uint8_t defaultExecuteFlags;
 
 }
 
-#endif // INSTMETADATA_H
+#endif