+++ test/API/OptionsTest.h
@@ -15,37 +15,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifndef INSTMETADATA_H
-#define INSTMETADATA_H
+#ifndef QBDITEST_OPTIONSTEST
+#define QBDITEST_OPTIONSTEST
 
-#include "llvm/MC/MCInst.h"
+#include <memory>
+#include "VM.h"
+#include "TestSetup/InMemoryAssembler.h"
 
-#include "Utility/InstAnalysis_prive.h"
+class OptionsTest {
+    protected:
+        OptionsTest() : vm() {}
 
-#include "State.h"
-
-namespace QBDI {
-
-class InstMetadata {
-public:
-    llvm::MCInst inst;
-    rword address;
-    uint32_t instSize;
-    uint32_t patchSize;
-    bool modifyPC;
-    bool merge;
-    uint8_t execblockFlags;
-    mutable InstAnalysisPtr analysis;
-
-    inline rword endAddress() const {
-        return address + instSize;
-    }
-
-    inline InstMetadata lightCopy() const {
-        return {inst, address, instSize, patchSize, modifyPC, merge, execblockFlags, nullptr};
-    }
+        QBDI::VM vm;
 };
 
-}
-
-#endif // INSTMETADATA_H
+#endif /* QBDITEST_OPTIONSTEST */